Student form implemented except Guardian data in form

+++ Development/Source Code/Sparkle/Crystal/Configuration/Crystal.Configuration.Component/Mobile/Data.cs
@@ -6,7 +6,7 @@ namespace Crystal.Configuration.Component.Mobile
     public class Data : BinAff.Core.Data
     {
 
-        public Int32 Number { get; set; }
+        public Int64 Number { get; set; }
 
         public Country.Data Country { get; set; }
